apps/tts-worker/requirements.txt
elevenlabs>=2.0.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.18.0
google-cloud-pubsub>=2.20.0
google-cloud-firestore>=2.14.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydub>=0.25.1
